edit product image

+++ projects/server/src/controllers/productController.js
@@ -88,26 +88,26 @@ module.exports = {
 
   async updateProductDetails(req, res) {
     const { id } = req.params;
-    const { name, price, weight, category_id, description } = req.body;
+    const editableFields = ['name', 'price', 'weight', 'category_id', 'description'];
     const t = await db.sequelize.transaction();
-
+  
     try {
       const product = await db.Product.findByPk(id, { transaction: t });
-
+  
       if (!product) {
         await t.rollback();
         return res.status(404).send({ message: "Product not found!" });
       }
-
-      product.name = name;
-      product.price = price;
-      product.weight = weight;
-      product.category_id = category_id;
-      product.description = description;
-
+  
+      editableFields.forEach(field => {
+        if (req.body[field] !== undefined) {
+          product[field] = req.body[field];
+        }
+      });
+  
       await product.save({ transaction: t });
       await t.commit();
-
+  
       return res.status(200).send({
         message: "Product updated successfully",
         data: product,
@@ -122,14 +122,12 @@ module.exports = {
     }
   },
 
-  async updateProductImage(req, res) {
+  async addImageToProduct(req, res) {
     const product_id = req.params.id;
-    const imgProductId = req.body.img_product_id;
     const newImageFile = req.file;
-
   
     const t = await db.sequelize.transaction();
-
+  
     try {
       const product = await db.Product.findByPk(product_id, { transaction: t });
       if (!product) {
@@ -137,29 +135,64 @@ module.exports = {
         return res.status(404).send({ message: "Product not found!" });
       }
   
+      const newImage = await db.Image_product.create({
+        product_id: product_id,
+        img_product: createProductImageDBPath(newImageFile.filename),
+        createdAt: new Date(),
+        updatedAt: new Date(),
+      }, { transaction: t });
+  
+      await t.commit();
+  
+      return res.status(200).send({
+        message: "Image added to product successfully",
+        data: newImage,
+      });
+    } catch (error) {
+      await t.rollback();
+      console.error("Error:", error);
+      return res.status(500).send({
+        message: "Fatal error on server",
+        errors: error.message,
+      });
+    }
+  },
+
+  async updateProductImage(req, res) {
+    const imgProductId = req.params.id;
+    const newImageFile = req.file;
+  
+    const t = await db.sequelize.transaction();
+  
+    try {
       const updatedImage = await db.Image_product.findByPk(imgProductId, { transaction: t });
       if (!updatedImage) {
         await t.rollback();
         return res.status(404).send({ message: "Image not found!" });
       }
   
-      if (updatedImage.img_product) {
-        const oldImageFilename = extractFilenameFromDBPath(updatedImage.img_product);
-        const oldImagePath = getAbsoluteProductImagePath(oldImageFilename);
+      if (newImageFile) {
+        if (updatedImage.img_product) {
+          const oldImageFilename = extractFilenameFromDBPath(updatedImage.img_product);
+          const oldImagePath = getAbsoluteProductImagePath(oldImageFilename);
   
-        try {
-          await fs.unlink(oldImagePath);
-        } catch (err) {
-          console.error("Error deleting old image file:", err);
+          try {
+            await fs.unlink(oldImagePath);
+          } catch (err) {
+            console.error("Error deleting old image file:", err);
+          }
         }
+  
+        updatedImage.img_product = createProductImageDBPath(newImageFile.filename);
       }
+  
       await updatedImage.save({ transaction: t });
   
       await t.commit();
-
+  
       return res.status(200).send({
         message: "Product image updated successfully",
-        data: product,
+        data: updatedImage, // Use the updatedImage variable instead of 'product'
       });
     } catch (error) {
       await t.rollback();
@@ -170,66 +203,19 @@ module.exports = {
       });
     }
   },
-
-  async addImagesToProduct(req, res) {
-    const product_id = req.params.id;
-    const newImageFiles = req.files;
   
-    const t = await db.sequelize.transaction();
-  
-    try {
-      const product = await db.Product.findByPk(product_id, { transaction: t });
-      if (!product) {
-        await t.rollback();
-        return res.status(404).send({ message: "Product not found!" });
-      }
-  
-      const newImages = await db.Image_product.bulkCreate(
-        newImageFiles.map((file) => ({
-          product_id: product_id,
-          img_product: createProductImageDBPath(file.filename),
-          createdAt: new Date(),
-          updatedAt: new Date(),
-        })),
-        { transaction: t }
-      );
-  
-      await t.commit();
-  
-      return res.status(200).send({
-        message: "Images added to product successfully",
-        data: newImages,
-      });
-    } catch (error) {
-      await t.rollback();
-      console.error("Error:", error);
-      return res.status(500).send({
-        message: "Fatal error on server",
-        errors: error.message,
-      });
-    }
-  },
-
   async deleteProductImage(req, res) {
-    const product_id = req.params.id;
-    const imgProductId = req.body.img_product_id;
+    const imgProductId = req.params.id;
   
     const t = await db.sequelize.transaction();
   
     try {
-      const product = await db.Product.findByPk(product_id, { transaction: t });
-      if (!product) {
-        await t.rollback();
-        return res.status(404).send({ message: "Product not found!" });
-      }
-  
       const deletedImage = await db.Image_product.findByPk(imgProductId, { transaction: t });
       if (!deletedImage) {
         await t.rollback();
         return res.status(404).send({ message: "Image not found!" });
       }
-      await deletedImage.destroy({ transaction: t });
-
+  
       if (deletedImage.img_product) {
         const deletedImageFilename = extractFilenameFromDBPath(deletedImage.img_product);
         const deletedImagePath = getAbsoluteProductImagePath(deletedImageFilename);
@@ -240,6 +226,8 @@ module.exports = {
           console.error("Error deleting image file:", err);
         }
       }
+  
+      await deletedImage.destroy({ transaction: t });
   
       await t.commit();
